refactored auth stuff in appjs into auth-hook
